LPD-29797 Extract fragment related external reference code upgraded processes to classes, so they can be tested
