+++ modules/apps/fragment/fragment-service/src/main/java/com/liferay/fragment/internal/upgrade/v2_11_1/FragmentCollectionExternalReferenceCodeUpgradeProcess.java
@@ -1,0 +1,21 @@
+/**
+ * SPDX-FileCopyrightText: (c) 2023 Liferay, Inc. https://liferay.com
+ * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
+ */
+
+package com.liferay.fragment.internal.upgrade.v2_11_1;
+
+import com.liferay.portal.kernel.upgrade.BaseExternalReferenceCodeUpgradeProcess;
+
+/**
+ * @author Rubén Pulido
+ */
+public class FragmentCollectionExternalReferenceCodeUpgradeProcess
+	extends BaseExternalReferenceCodeUpgradeProcess {
+
+	@Override
+	protected String[][] getTableAndPrimaryKeyColumnNames() {
+		return new String[][] {{"FragmentCollection", "fragmentCollectionId"}};
+	}
+
+}
